+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Karafka framework changelog
 
+## 2.0.0-beta5 (Unreleased)
+- Always resume processing of a revoked partition just in case it gets re-assigned.
+- Improve specs stability
+
 ## 2.0.0-beta4 (2022-06-20)
 - Rename job internal api methods from `#prepare` to `#before_call` and from `#teardown` to `#after_call` to abstract away jobs execution from any type of executors and consumers logic
 - Remove ability of running `before_consume` and `after_consume` completely. Those should be for internal usage only.